mentorAulaVirtual/requirements.txt
asgiref==3.7.2
black==23.7.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
crispy-bootstrap5==0.7
Django==4.2.2
django-crispy-forms==2.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
docopt==0.6.2
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
phonenumbers==8.13.15
Pillow==10.0.0
platformdirs==3.9.1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
yarg==0.1.9
